Added the renderer module
