+++ index.html
@@ -28,7 +28,7 @@
 
             <form class="task form__input" id="taskInput">
                 <label for="new_task" class="checker" data-todo_index="${index}"></label>
-                <input type="text" id="new_task" name="new_task" autocomplete="off" />
+                <input type="text" id="new_task" name="new_task" autocomplete="off" required="true"/>
                 <span class="platform">Created on web</span>
             </form>
         </section>
